+++ unicorn_binance_websocket_api/unicorn_binance_websocket_api_manager.py
@@ -266,7 +266,10 @@ class BinanceWebSocketApiManager(threading.Thread):
                                    api_key=False,
                                    api_secret=False,
                                    symbols=False,
-                                   output=False):
+                                   output=False,
+                                   ping_interval=False,
+                                   ping_timeout=False,
+                                   close_timeout=False):
         """
         Create a list entry for new streams
 
@@ -296,9 +299,28 @@ class BinanceWebSocketApiManager(threading.Thread):
                        convert with `UnicornFy <https://github.com/oliver-zehentleitner/unicorn_fy>`_ -  otherwise with
                        the default setting "raw_data" the output remains unchanged and gets delivered as received from
                        the endpoints
+        :param ping_interval: Once the connection is open, a `Ping frame`_ is sent every
+                              `ping_interval` seconds. This serves as a keepalive. It helps keeping
+                              the connection open, especially in the presence of proxies with short
+                              timeouts on inactive connections. Set `ping_interval` to `None` to
+                              disable this behavior. (default: 20)
+                              This parameter is passed through to the `websockets.client.connect()
+                              <https://websockets.readthedocs.io/en/stable/api.html?highlight=ping_interval#websockets.client.connect>_
+        :type ping_interval: int or None
+        :param ping_timeout: If the corresponding `Pong frame` isn't received within
+                             `ping_timeout` seconds, the connection is considered unusable and is closed with
+                             code 1011. This ensures that the remote endpoint remains responsive. Set
+                             `ping_timeout` to `None` to disable this behavior. (default: 20)
+                             This parameter is passed through to the `websockets.client.connect()
+                             <https://websockets.readthedocs.io/en/stable/api.html?highlight=ping_interval#websockets.client.connect>_
+        :type ping_timeout: int or None
+        :param close_timeout: The `close_timeout` parameter defines a maximum wait time in seconds for
+                              completing the closing handshake and terminating the TCP connection. (default: 10)
+                              This parameter is passed through to the `websockets.client.connect()
+                              <https://websockets.readthedocs.io/en/stable/api.html?highlight=ping_interval#websockets.client.connect>_
+        :type close_timeout: int or None
         :type output: str
         """
-
         if api_key is not False and api_secret is not False:
             add_api_key = api_key
             add_api_secret = api_secret
@@ -323,6 +345,9 @@ class BinanceWebSocketApiManager(threading.Thread):
                                        'api_key': copy.deepcopy(add_api_key),
                                        'api_secret': copy.deepcopy(add_api_secret),
                                        'dex_user_address': copy.deepcopy(self.dex_user_address),
+                                       'ping_interval': copy.deepcopy(ping_interval),
+                                       'ping_timeout': copy.deepcopy(ping_timeout),
+                                       'close_timeout': copy.deepcopy(close_timeout),
                                        'status': 'starting',
                                        'start_time': time.time(),
                                        'processed_receives_total': 0,
@@ -872,7 +897,10 @@ class BinanceWebSocketApiManager(threading.Thread):
                       api_key=False,
                       api_secret=False,
                       symbols=False,
-                      output=False):
+                      output=False,
+                      ping_interval=20,
+                      ping_timeout=20,
+                      close_timeout=10):
         """
         Create a websocket stream
 
@@ -943,6 +971,26 @@ class BinanceWebSocketApiManager(threading.Thread):
                        the default setting "raw_data" the output remains unchanged and gets delivered as received from
                        the endpoints
         :type output: str
+        :param ping_interval: Once the connection is open, a `Ping frame`_ is sent every
+                              `ping_interval` seconds. This serves as a keepalive. It helps keeping
+                              the connection open, especially in the presence of proxies with short
+                              timeouts on inactive connections. Set `ping_interval` to `None` to
+                              disable this behavior. (default: 20)
+                              This parameter is passed through to the `websockets.client.connect()
+                              <https://websockets.readthedocs.io/en/stable/api.html?highlight=ping_interval#websockets.client.connect>_
+        :type ping_interval: int or None
+        :param ping_timeout: If the corresponding `Pong frame` isn't received within
+                             `ping_timeout` seconds, the connection is considered unusable and is closed with
+                             code 1011. This ensures that the remote endpoint remains responsive. Set
+                             `ping_timeout` to `None` to disable this behavior. (default: 20)
+                             This parameter is passed through to the `websockets.client.connect()
+                             <https://websockets.readthedocs.io/en/stable/api.html?highlight=ping_interval#websockets.client.connect>_
+        :type ping_timeout: int or None
+        :param close_timeout: The `close_timeout` parameter defines a maximum wait time in seconds for
+                              completing the closing handshake and terminating the TCP connection. (default: 10)
+                              This parameter is passed through to the `websockets.client.connect()
+                              <https://websockets.readthedocs.io/en/stable/api.html?highlight=ping_interval#websockets.client.connect>_
+        :type close_timeout: int or None
         :return: stream_id or 'False'
         """
         # create a stream
@@ -992,7 +1040,10 @@ class BinanceWebSocketApiManager(threading.Thread):
                                         symbols=symbols,
                                         api_key=api_key,
                                         api_secret=api_secret,
-                                        output=output)
+                                        output=output,
+                                        ping_interval=ping_interval,
+                                        ping_timeout=ping_timeout,
+                                        close_timeout=close_timeout)
         try:
             loop = asyncio.new_event_loop()
         except OSError as error_msg: